pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "ghreport"
authors = [{name = "Graham Wheeler", email = "graham@grahamwheeler.com"}]
readme = "README.md"
license = {file = "LICENSE"}
dynamic = ["version", "description"]
dependencies = [
    "flit ==3.10.1",
    "numpy == 2.2.2",
    "click==8.1.8",
    "pygithub==2.5.0",
    "gidgethub==5.3.0",
    "httpx==0.28.1",
    "pytz==2025.1",
    "matplotlib == 3.10.0",
    "seaborn==0.13.2",
    "wordcloud==1.9.4",
]
classifiers = ["Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.optional-dependencies]
dashboard = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "aiosqlite>=0.20",
    "apscheduler>=3.10,<4",
    "pydantic-settings>=2.2",
]

[project.urls]
Source = "https://github.com/gramster/ghreport"

[project.scripts]
ghreport = "ghreport:main"
ghreport-dashboard = "ghreport.dashboard.app:main"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
